Add comments, create room prototype

+++ Source/UEPlugin_ISReverb/ReflectorEdge.cpp
@@ -9,12 +9,14 @@ ReflectorEdge::ReflectorEdge(FVector3f a, FVector3f b)
 	PointB = b;
 }
 
+// Returns unit vector expressing the direction followed by the edge from pointA to pointB
 FVector3f ReflectorEdge::Direction()
 {
 	return (PointA - PointB).GetSafeNormal();
 }
-	
-double ReflectorEdge::Length()
+
+// Returns length of the edge in units
+float ReflectorEdge::Length()
 {
 	return (PointB - PointA).Length();
 }
